+++ pkg/flow/data/instance_event.go
@@ -12,10 +12,10 @@
  * Unless required by applicable law or agreed to in writing, software distributed
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissionsr and limitations under the License.
+ * specific language governing permissions and limitations under the License.
  */
 
-package utils
+package data
 
 import (
 	"github.com/polarismesh/polaris-go/pkg/model"
